BSM-30, BSM-31: Implement RCA and IA algorithms.

+++ features/bsm/service/api/src/main/java/org/opennms/netmgt/bsm/service/model/graph/BusinessServiceGraph.java
@@ -40,7 +40,7 @@ public interface BusinessServiceGraph extends DirectedGraph<GraphVertex,GraphEdg
 
     GraphVertex getVertexByReductionKey(String reductionKey);
 
-    GraphVertex getVertexByIpServiceId(Long id);
+    GraphVertex getVertexByIpServiceId(Integer id);
 
     GraphVertex getVertexByEdgeId(Long id);
